backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.9.0.80
numpy==1.24.3
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
gfpgan==1.3.8
basicsr==1.4.2
facexlib==0.3.0
realesrgan==0.3.0
tb-nightly==2.16.0a20240119
